services/packing-assistant/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
google-generativeai==0.4.1
packaging
